fixed bug where gui couldn't save numpy floats to md file. added chunk size limit to minimize crashing the computer by requesting too many resources

+++ territorytools/gui.py
@@ -991,7 +991,7 @@ class SlideInputer(QGroupBox):
 
     def set_value(self, value):
         if self.dtype == 'float':
-            value = np.round(value, 2)
+            value = float(np.round(value, 2))
         self.cur_val = value
         self.ebox.setText(str(value))
         self.slide.setValue(np.round(((self.cur_val - self.offset) * self.scale)).astype(int))
